requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
openai==0.28
PyPDF2
pillow